# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Angola Viva",
        "description": "Documentação da API do sistema Angola Viva",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/relatorio": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Listar relatórios",
                "operationId": "c150ca7c99915643decfcd55c59812a7",
                "responses": {
                    "200": {
                        "description": "Lista de relatórios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao listar relatórios"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Criar relatório",
                "operationId": "863d7307477ec019cc689c6ac1b0489e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "dados"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "casos_por_regiao",
                                            "evolucao_temporal",
                                            "distribuicao_demografica",
                                            "casos_por_sexo",
                                            "casos_por_idade",
                                            "casos_por_hospital",
                                            "casos_por_municipio",
                                            "casos_por_resultado_triagem",
                                            "outro"
                                        ]
                                    },
                                    "dados": {
                                        "description": "JSON dos dados do relatório",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relatório criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Não é possível salvar relatório sem dados."
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro ao criar relatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/relatorio/{id}": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Exibir relatório",
                "operationId": "90dfa7c2cbe81c582b472907118660f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do relatório",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do relatório",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relatório não encontrado"
                    },
                    "500": {
                        "description": "Erro ao exibir relatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Excluir relatório",
                "operationId": "2a53c7fa6a9d45c0506d4f781d1ee1b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do relatório",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relatório removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relatório não encontrado"
                    },
                    "500": {
                        "description": "Erro ao remover relatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/relatorio/{id}/pdf": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Gerar PDF do relatório",
                "operationId": "b1513555f5abed7c3fb74c74ac27d65f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do relatório",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF gerado (base64)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "file": {
                                            "description": "PDF em base64",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relatório não encontrado"
                    },
                    "500": {
                        "description": "Erro ao gerar PDF"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilize o token JWT no formato Bearer para autenticação.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}